Made CSS Desktop By Default
The initial commit had the website styled as a mobile site and would change as the width of the browser got wider. This has been reversed so the website is styled as a desktop site and changes as the width of the browser gets smaller. This is because this website is meant to be viewed on a desktop, so it makes more sense to develop for that platform first.