Updated the Links on the Navbar
Changed the navbar links to "Attendance" and "My Account". "Attendance" will detect the user's account type based on the Session data that will be stored and will redirect the user to the student or faculty attendance pages depending on account type. The "My Account" link does not work on this branch, but should once it merges with the testing branch. It just goes to MyTesting.cshtml .