Added Login Without DB Connection
Created methods to attempt logging into a user account. One method, TempAttemptLogin, is temporary to see how the functionality will work. Any account works and the password is "Password". The other method, AttemptLogin, will be used in the full release once it is connected to the account data in the database. Regardless of the method, session variables are created to track the email and account type, and the sessions currently expire after 1 minute of inactivity for testing. Updated the navbar to display Login/Register if the user is not signed in and Attendance/My Account if they are signed in. Removed parameters from some methods because they can use session data instead. Added a new style to make forms look nicer.