Basic Registration Complete
FINALLY got the registration done.
This has password and email validation checks.
It is "basic" because it is only part 1, but without any of the email verification included, which makes up the finale of part 1 and all of part 2. Part 3 is not included becaise it will also use future email creation and sending knowledge.