Filter Student Attendance by Time Frame
A time frame id number can be entered into the StudentAttendance method. If no time frame is given, it defaults to the current time frame. If a time frame is not currently happening, it will go to the time frame with the latest start date. If there is no time frame with the latest start date (implying that there are no time frames), the StudentAttendance method returns NotFound(). Depending on the time frame, the specified user's attended events are shown.